backend/requirement.txt
flask
flask_cors
python-dotenv
google-genai
requests
beautifulsoup4
tqdm